Modified: ErrorBoundary status output; npm build; Refactored: Routes && Redux into States
