+++ src/Components/Authentication/SignInWithGoogleFirebase/index.tsx
@@ -3,7 +3,7 @@ import { useDispatch } from "react-redux";
 import {
   googleSignInStart,
   emailSignInStart,
-} from "../../../Redux/Users/user.action";
+} from "../../../States/Redux/Users/user.action";
 
 import LockOutlinedIcon from "@mui/icons-material/LockOutlined";
 import { createTheme, ThemeProvider } from "@mui/material/styles";
