--- src/Redux/CombineReducers/index.ts
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-
-import { userReducer } from '../Users/user.reducer';
-
-export const rootReducer = combineReducers({
-  user: userReducer,
-});
