+++ src/States/Redux/CombineReducers/index.ts
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+
+import { userReducer } from '../Users/user.reducer';
+
+export const rootReducer = combineReducers({
+  user: userReducer,
+});
