+++ src/States/Redux/Store/index.ts
@@ -1,0 +1,34 @@
+import { compose, createStore, applyMiddleware, Middleware } from 'redux';
+import { persistStore } from 'redux-persist';
+import logger from 'redux-logger';
+import createSagaMiddleware from 'redux-saga';
+import { rootSaga } from '../Root-Saga';
+import { rootReducer } from '../CombineReducers';
+export type RootState = ReturnType<typeof rootReducer>;
+
+declare global {
+  interface Window {
+    __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: typeof compose;
+  }
+}
+
+const sagaMiddleware = createSagaMiddleware();
+
+const middleWares = [
+  process.env.NODE_ENV !== 'production' && logger,
+  sagaMiddleware,
+].filter((middleware): middleware is Middleware => Boolean(middleware));
+
+const composeEnhancer =
+  (process.env.NODE_ENV !== 'production' &&
+    window &&
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) ||
+  compose;
+
+const composedEnhancers = composeEnhancer(applyMiddleware(...middleWares));
+
+export const store = createStore(rootReducer, undefined, composedEnhancers)
+
+sagaMiddleware.run(rootSaga);
+
+export const persistor = persistStore(store);
