+++ src/States/Redux/Users/user.selector.ts
@@ -1,0 +1,9 @@
+import { createSelector } from 'reselect';
+import { RootState } from '../Store';
+import { UserState } from './user.reducer';
+
+export const selectUserReducer = (state: RootState): UserState => state.user;
+export const selectCurrentUser = createSelector(
+  selectUserReducer,
+  (user) => user.currentUser
+);
